Integrate CPUMonitor Class in monitor.py and Import in JobRunner
